+++ cacheman/lib/cacheman.dart
@@ -1,5 +1,3 @@
-library cacheman;
-
 export 'src/cacheman.dart';
 export 'src/debug.dart';
 export 'src/engine.dart' show CacheOptions, CachemanOptions, Engine;
@@ -8,4 +6,3 @@ export 'src/fast.dart';
 export 'src/interface.dart' show Store, MemoCache, Codec, CachemanOnError;
 export 'src/jsonx.dart' show Jsonx;
 export 'src/memo.dart' show Memo;
-export 'src/memory.dart' show Memory;